requirements.txt
Flask
Flask-WTF
WTForms
icalendar
qrcode[pil]
Pillow
pytz
gunicorn
